+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/BackendInsight.story.tsx
@@ -10,11 +10,11 @@ import { LINE_CHART_CONTENT_MOCK, LINE_CHART_CONTENT_MOCK_EMPTY } from '../../..
 import { CodeInsightsBackendContext } from '../../../../core/backend/code-insights-backend-context'
 import { CodeInsightsSettingsCascadeBackend } from '../../../../core/backend/setting-based-api/code-insights-setting-cascade-backend'
 import { InsightInProcessError } from '../../../../core/backend/utils/errors'
-import { InsightExecutionType, InsightType } from '../../../../core/types'
+import { BackendInsight, InsightExecutionType, InsightType, isCaptureGroupInsight } from '../../../../core/types'
 import { SearchBackendBasedInsight } from '../../../../core/types/insight/search-insight'
 import { SETTINGS_CASCADE_MOCK } from '../../../../mocks/settings-cascade'
 
-import { BackendInsight } from './BackendInsight'
+import { BackendInsightView } from './BackendInsight'
 
 const { add } = storiesOf('web/insights/BackendInsight', module).addDecorator(story => (
     <WebStory>{() => story()}</WebStory>
@@ -27,6 +27,7 @@ const INSIGHT_CONFIGURATION_MOCK: SearchBackendBasedInsight = {
     type: InsightExecutionType.Backend,
     viewType: InsightType.SearchBased,
     id: 'searchInsights.insight.mock_backend_insight_id',
+    step: { weeks: 2 },
 }
 
 const mockInsightAPI = ({
@@ -36,7 +37,11 @@ const mockInsightAPI = ({
     hasData = true,
 } = {}) => {
     class CodeInsightsStoryBackend extends CodeInsightsSettingsCascadeBackend {
-        public getBackendInsightData = (insight: SearchBackendBasedInsight) => {
+        public getBackendInsightData = (insight: BackendInsight) => {
+            if (isCaptureGroupInsight(insight)) {
+                throw new Error('This demo does not support capture group insight')
+            }
+
             if (throwProcessingError) {
                 return throwError(new InsightInProcessError())
             }
@@ -57,7 +62,7 @@ const mockInsightAPI = ({
 }
 
 const TestBackendInsight: React.FunctionComponent = () => (
-    <BackendInsight
+    <BackendInsightView
         style={{ width: 400, height: 400 }}
         insight={INSIGHT_CONFIGURATION_MOCK}
         telemetryService={NOOP_TELEMETRY_SERVICE}