+++ enterprise/cmd/frontend/internal/codeintel/httpapi/upload_handler.go
@@ -13,19 +13,19 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/stores/uploadstore"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
+	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 )
 
 type UploadHandler struct {
-	db          dbutil.DB
+	db          database.DB
 	dbStore     DBStore
 	uploadStore uploadstore.Store
 	operations  *Operations
 }
 
 func NewUploadHandler(
-	db dbutil.DB,
+	db database.DB,
 	dbStore DBStore,
 	uploadStore uploadstore.Store,
 	internal bool,